[14.0][IMP] Send failed message with sudo (to ensure send to record user)
If job fails, then message is send to users. But, as message is send with record user, then this message is not send by email to this user (that is the most important).

Please review that, I can not tested, but with sudo it must be send as admin user.